Added changes proposed in code review
