+++ components/inspectit-ocelot-configurationserver/src/main/java/rocks/inspectit/ocelot/file/FileManager.java
@@ -2,12 +2,10 @@ package rocks.inspectit.ocelot.file;
 
 import com.google.common.annotations.VisibleForTesting;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.io.FileUtils;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
 import rocks.inspectit.ocelot.config.model.InspectitServerSettings;
 import rocks.inspectit.ocelot.file.dirmanagers.GitDirectoryManager;
 import rocks.inspectit.ocelot.file.dirmanagers.WorkingDirectoryManager;
@@ -16,7 +14,10 @@ import javax.annotation.PostConstruct;
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.*;
+import java.nio.file.AccessDeniedException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
 
 /**
@@ -79,9 +80,6 @@ public class FileManager {
      * @return A list of files found in the given path.
      */
     public List<FileInfo> listConfigurationFiles(String path, boolean recursive, boolean versioning) throws IOException {
-        if (!StringUtils.isEmpty(path)) {
-            assertPathWithinFilesRoot(path);
-        }
         if (versioning) {
             return gitDirectoryManager.listFiles(setToPathInConfig(path), recursive);
         } else {
@@ -98,9 +96,6 @@ public class FileManager {
      * @return A list of files found in the given path.
      */
     public List<FileInfo> listSpecialFiles(String path, boolean recursive, boolean versioning) throws IOException {
-        if (!StringUtils.isEmpty(path)) {
-            assertPathWithinFilesRoot(path);
-        }
         if (versioning) {
             return gitDirectoryManager.listFiles(path, recursive);
         } else {
@@ -108,18 +103,13 @@ public class FileManager {
         }
     }
 
-   /* public List<FileInfo> listFiles(String path, boolean recursive) throws IOException {
-        return workingDirectoryManager.listFiles(path, recursive);
-    }*/
-
     /**
      * @param path the path to check
      * @return true if the given path denotes a directory
      * @throws AccessDeniedException if access is forbidden
      */
     public boolean isDirectory(String path) throws AccessDeniedException {
-        assertPathWithinFilesRoot(path);
-        return Files.isDirectory(filesRoot.resolve(path));
+        return workingDirectoryManager.isDirectory(path);
     }
 
     /**
@@ -128,8 +118,7 @@ public class FileManager {
      * @throws AccessDeniedException if access is forbidden
      */
     public boolean exists(String path) throws AccessDeniedException {
-        assertPathWithinFilesRoot(path);
-        return Files.exists(filesRoot.resolve(path));
+        return workingDirectoryManager.exists(path);
     }
 
     /**
@@ -139,10 +128,8 @@ public class FileManager {
      * @throws IOException if the directory already exists or could not be created for any reason
      */
     public synchronized void createDirectory(String path) throws IOException, GitAPIException {
-        assertValidSubPath(path);
-        Path dir = filesRoot.resolve(path);
-
-        FileUtils.forceMkdir(dir.toFile());
+        workingDirectoryManager.createDirectory(path);
+        commitAllChanges();
         fireFileChangeEvent();
     }
 
@@ -153,14 +140,7 @@ public class FileManager {
      * @throws IOException if the directory could not be deleted
      */
     public synchronized void deleteDirectory(String path) throws IOException, GitAPIException {
-        assertValidSubPath(path);
-        Path dir = filesRoot.resolve(path);
-        // throw a more meaningful exception instead of the illegal argument exception thrown by
-        // FileUtils.deleteDirectory
-        if (!Files.exists(dir) || !Files.isDirectory(dir)) {
-            throw new NotDirectoryException(getRelativePath(dir));
-        }
-        FileUtils.deleteDirectory(dir.toFile());
+        workingDirectoryManager.deleteDirectory(path);
         commitAllChanges();
         fireFileChangeEvent();
     }
@@ -174,7 +154,6 @@ public class FileManager {
      * @throws IOException if the file could not be read.
      */
     public String readConfigurationFile(String path, boolean versioning) throws AccessDeniedException {
-        assertValidSubPath(path);
         if (!versioning) {
             try {
                 return workingDirectoryManager.readFile(setToPathInConfig(path));
@@ -201,7 +180,6 @@ public class FileManager {
      * @throws IOException if the file could not be read.
      */
     public String readSpecialFile(String path, boolean versioning) throws AccessDeniedException {
-        assertValidSubPath(path);
         if (!versioning) {
             try {
                 return workingDirectoryManager.readFile(path);
@@ -229,7 +207,6 @@ public class FileManager {
      * @throws IOException if the file could not be written
      */
     public synchronized void writeConfigurationFile(String path, String content) throws IOException, GitAPIException {
-        assertValidSubPath(path);
         workingDirectoryManager.writeFile(setToPathInConfig(path), content);
         commitAllChanges();
         fireFileChangeEvent();
@@ -256,15 +233,10 @@ public class FileManager {
      * @param path the path of the file to delete
      * @throws IOException if the file could not be deleted.
      */
-    public synchronized void deleteFile(String path) throws IOException {
-        assertValidSubPath(path);
-        Path file = filesRoot.resolve(path);
-        if (Files.isRegularFile(file)) {
-            Files.delete(file);
-            fireFileChangeEvent();
-        } else {
-            throw new AccessDeniedException(path);
-        }
+    public synchronized void deleteFile(String path) throws IOException, GitAPIException {
+        workingDirectoryManager.deleteFile(path);
+        fireFileChangeEvent();
+        commitAllChanges();
     }
 
     /**
@@ -276,8 +248,6 @@ public class FileManager {
      * @throws IOException if the given file or directory could not be renamed / moved
      */
     public synchronized void move(String source, String destination) throws IOException, GitAPIException {
-        assertValidSubPath(source);
-        assertValidSubPath(destination);
         workingDirectoryManager.move(source, destination);
         commitAllChanges();
     }
@@ -290,47 +260,6 @@ public class FileManager {
         return filesRoot.relativize(f.normalize())
                 .toString()
                 .replace(f.getFileSystem().getSeparator(), "/");
-    }
-
-    /**
-     * Ensures that the given path is a subpath of {@link #filesRoot}
-     *
-     * @param path the pat hto check
-     * @throws AccessDeniedException if the file is not a subpath of the filesRoot
-     */
-    private void assertValidSubPath(String path) throws AccessDeniedException {
-        assertPathNotEmpty(path);
-        assertPathWithinFilesRoot(path);
-    }
-
-    /**
-     * Ensures that the given path does not point to the filesRoot directory.
-     *
-     * @param path the path to test
-     * @throws AccessDeniedException thrown if the path points to the filesRoot
-     */
-    private void assertPathNotEmpty(String path) throws AccessDeniedException {
-        if (StringUtils.isEmpty(path)) {
-            throw new AccessDeniedException("/");
-        }
-        if (filesRoot.resolve(path).toAbsolutePath().normalize().equals(filesRoot.toAbsolutePath())) {
-            throw new AccessDeniedException("/");
-        }
-    }
-
-    /**
-     * Ensures that the given path does not point to a file outside of the filesRoot directory.
-     * This method succeeds if the given path represents the filesRoot.
-     *
-     * @param path the path to test
-     * @throws AccessDeniedException thrown if the path points to a file outside of the filesRoot
-     */
-    private void assertPathWithinFilesRoot(String path) throws AccessDeniedException {
-        String subPath = filesRoot.resolve(path).toAbsolutePath().normalize().toString();
-        String rootPath = filesRoot.toString();
-        if (!subPath.startsWith(rootPath)) {
-            throw new AccessDeniedException(path);
-        }
     }
 
     /**
